refactor: rename the Intercom window to Chat (#293)
